msm: rmt_storage: Do not unregister client if there are pending requests
As seen in a corner case where the MODEM initiates an efs_sync from FS
task and waits for the response from APPS. At the same time, APPS reboot
was triggered which causes remoteFS to send force_sync request to
MODEM. Now APPS polls for sync_status from MODEM. Since an efs_sync is
ongoing, there is no more dirty-data to write and MODEM sends success
response to APPS. This causes remoteFS on APPS to proceed with
unregistering the client while the ongoing efs_sync is not completed,
causing MODEM to dog timeout on FS task.

To avoid such cases, remoteFS should wait to check for any ongoing
requests prior to unregistering.

CRs-Fixed: 435986
Change-Id: Ia12263d17099bfc928ed9ae93114050b986e69cc
Signed-off-by: Pratibhasagar V <pratibha@codeaurora.org>