sweep2wake: Add sensitive option
This is for users (like myself) who feel like the current s2w is too tedious/precise to trigger. This option, when enabled only requires the 1st and last button to be triggered making it easier to initiate with a thumb or without making full contact & discontact with all 4 buttons.

Conflicts:
	drivers/input/keyboard/cypress-touchkey.c

Conflicts:
	drivers/input/keyboard/cypress-touchkey.c
	drivers/input/keyboard/s2w-switch.c
	include/linux/s2w-switch.h

Change-Id: I8bbdf0675c01e7ef57c555e77333915a4e59e44f