sweep2wake: Add sensitive option
This is for users (like myself) who feel like the current s2w is too tedious/precise to trigger. This option, when enabled only requires the 1st and last button to be triggered making it easier to initiate with a thumb or without making full contact & discontact with all 4 buttons.

Change-Id: I39643f07e67c9aeeda7341c912d97307bc2e9d6d